+++ opamp/opentelemetry-opamp-client/CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Add support for reporting non-JSON effective config bodies.
+
 ## Version 0.2b0 (2026-04-01)
 
 - Breaking change: callback class `Callbacks` renamed to `OpAMPCallbacks`